reorg components
