+++ components/dashboard/DashboardContent/PasswordTable/TableEntry/ClosedEntry.tsx
