+++ components/dashboard/DashboardContent/PasswordTable/TableEntry/ExpandedEntry.tsx
