+++ components/dashboard/DashboardContent/PasswordTable/TableEntry/ExpandedEntry/DeleteModal.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../../styles/colors.module";
+@use "../../../../../../styles/colors.module";
 $line_height: 100px;
 
 .modal_body {
