+++ components/dashboard/DashboardContent/PasswordTable/TableEntry/ExpandedEntry/EntryInput.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../../styles/colors.module";
+@use "../../../../../../styles/colors.module";
 @use "../../TableEntry.module" as table_entry;
 
 .container {
