+++ components/dashboard/DashboardContent/SidePanel.module.scss
